Structure to support pyOCD.
